refactor: break down events mod.rs into focused modules
- Split 459-line mod.rs into organized module hierarchy
- Main mod.rs now only 65 lines, all new modules under 200 lines
- Group by function: types/, emitter/, global/, subscriber/
- Use directory hierarchies, no underscores
- Maintain all functionality and 18 tests passing
